cleanup after the debugging and getting integration testing working on GH
